+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## v3.3.0
+### Fixed
+- Fix regular expressions
+
+### Changed
+- Don't allow users to reenable while in queue
+
 ## v3.2.0
 ### Changed
 - Remove Alpine >= 3.13 restriction